Removed Character class, moved variables to Player
The Character class was pretty useless, so everything there was moved
to the Player class. There’s still a lot to be added and fixed, but I
thought I had to commit this because I also switched to the Xcode
environment… screw Eclipse not being able to recognize the standard
library.